msm: camera: Fix underflow in bus bandwidth calculation
When there is a total of 1 stream active, the bus bandwidth calculation would often underflow for both AB and IB vectors, causing very large and invalid bandwidth rates to be requested (ex. 18446744073209556616 bps). The absurd bandwidth requests would be ignored, quickly causing lethal bus overflow.

Use the correct bandwidth calculations to fix the underflow and vote for the correct bandwidth.