ASoC: wcd9xxx-mbhc: Don't sleep for first poll in correct-plug worker
There is a good chunk of time between when the correct-plug worker is scheduled and when the first poll inside the correct-plug worker begins, so no delay is necessary for the first poll inside the correct-plug worker.

Removing the delay on the first poll inside the correct-plug worker reduces detection latency by 100ms (HS_DETECT_PLUG_INERVAL_MS) when the correct-plug worker is necessary.

Signed-off-by: Sultanxda <sultanxda@gmail.com>