ASoC: wcd9xxx-mbhc: Introduce fast detection for audio jack
Our audio jack hardware is rather cheap, and accurate detections can take up
to 2 seconds for certain headphones or headsets. Previously, we would set the
detected plug-type to headphones or a headset if that was what the plug was
determined to be immediately, and then run the correct-plug worker afterwards
to ensure consistency and prevent an incorrect plug-type from being kept if our
assumption from the immediate detection was incorrect.

However, the detected plug type can fluctuate greatly while the correct-plug worker
is running, causing the detected plug to continuously switch. Constantly switching
the detected plug can have adverse effects on userspace. For example, plugging in a
pair of headphones would cause a running soundtrack to suddenly pause, due to the
false, constant plug insertions/removals being sent to Android.

With fast-detection, if a valid plug is detected immediately upon insertion, the
correct-plug worker will not constantly change the inserted plug type while it is
running, resulting in a quick, accurate detection to be seen by the user. The final
verdict from the correct-plug worker will be checked against what was found using
fast-detection, and if the result varies from what was determined using
fast-detection, then the correct-plug worker's result will take precedence and the
inserted plug type will be changed to that.

Overall, fast-detection provides a safe method to quickly register various headsets
and headphones, improving UX.

ASoC: wcd9xxx-mbhc: Ignore button interrupts during plug detection

If an inserted plug is mistakenly detected as a headset when it is not, it
can induce spurious button interrupts that interfere with the end user.

Ignore all button interrupts until plug-type detection is completely finished
in order to prevent havoc caused by spurious button interrupts.

ASoC: wcd9xxx-mbhc: Reduce detection time for PLUG_TYPE_HEADSET

In lucky cases where a good headset is plugged in, detection latency can be
significantly reduced by setting the headset plug type early on, before entering
the correct-plug worker.

If there is a faulty detection and an inserted plug is detected as a headset early-on,
even though it is not, we will just fall back to the correct-plug worker to fix the detection error.